Prevent false positives around WARN_ON()s and BUG_ON()s
Claude would find a WARN_ON(foo) and decide that foo was something that
could happen, and then ask whe we didn't check for foo in more places.

This commit shifts things to assume that WARN_ON() means we've decided
that condition doesn't need to be handled elsewhere in the code.  It
might miss some crashes etc but it better matches how we deal with
corner cases in kernel code.

Signed-off-by: Chris Mason <clm@meta.com>