Replace the pattern tables with markdown sections
The tables were hard to edit, especially for complex patterns, and they
weren't actually markdown, so they didn't render well.  This commit
swaps them out for header sections (usually level 4) and the result
is a lot easier to read.

The actual content of the reviews is mostly unchanged, +/- some editing
for clarity

Signed-off-by: Chris Mason <clm@fb.com>