add scripts/find_descendants.py
This is a helper to find all the commits that can reach your current
commit in the repository.  When we're reviewing patches, we often
use temporary branches without names, so we need a somewhat brute force
way to find our children.  This can be used in scripts to pass
the end point of a series being reviewed to the prompts.

Signed-off-by: Chris Mason <clm@meta.com>