Improve NULL pointer analysis
There were a few different problems with how claude was identifying
NULL pointer bugs:

It didn't actually understand what dereferencing a pointer meant

  - it thought foo->bar would dereference bar
  - added patterns/null.md to explain this

The kernel has a lot of "this pointer can't be NULL if xyz" implicit
knowledge scattered around, and it's pretty hard to understand if a
NULL check is actually required.

  - added patterns/guards.md to search for these implicit NULL guards

claude likes to wander off as the review drags on, making it less
likely to follow instructions properly during the false positive
elimination.

  - this seems to be unfixable, but I tried adjusting the instructions

This patch from Jonas Gorski was a great example of a complicated NULL
pointer deref false positive.  We still get this one wrong as percentage
of the time, but this commit is a big improvement.

https://lore.kernel.org/netdev/CAOiHx=mog+8Grm1QTnqU_F3=BnWmJqTj+ko-nZiRMAb4-hvSqw@mail.gmail.com/

Signed-off-by: Chris Mason <clm@meta.com>